added build definition for cpu fp16 ops and separated avx calls to another file
